前台注册type,类型问题bug(已解决)

+++ src/main/java/com/review/www/service/UserService.java
@@ -47,7 +47,7 @@ public interface UserService {
      * @param newPwd
      * @param type
      */
-    Result changePassword(String number, String newPwd, Byte type);
+    Result changePassword(String number, String newPwd, int type);
 
     /**
      * getByType
